Remove ACCEPT rule for OVN-to-localhost traffic
With iptables-based firewalls, you can subvert the firewall by adding
"-j ACCEPT" rules before the firewall rules. With nftables-based
firewalls, it's impossible to guarantee that your own rules will run
before the firewall's rules, so this won't reliably work; if you want
the firewall to not block your traffic, you need to actually configure
the firewall to not block your traffic.

(We still install ACCEPT rules in the FORWARD chain, because those are
needed as a result of *our own* DROP rule, when DisableForwarding is
set.)

Signed-off-by: Dan Winship <danwinship@redhat.com>